requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary # For PostgreSQL connection
Flask-Login
python-dotenv
pywebpush
gunicorn
pytest